requirements.txt
flask==2.2.5
Jinja2==3.1.2
itsdangerous==2.1.2
Werkzeug>=2.2,<2.3
click==8.1.7
python-dotenv==1.0.0
dnspython==2.7.0
pywebview==5.4
pyinstaller==6.13.0
openpyxl==3.1.2
reportlab==4.0.4
pytest==8.2.0
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.1.1